添加 CQ:xml的支持
