+++ coolq/api.go
@@ -622,101 +622,3 @@ func convertGroupMemberInfo(groupId int64, m *client.GroupMemberInfo) MSG {
 		"card_changeable":   false,
 	}
 }
-
-func (bot *CQBot) Send_private_msg_xml(userId int64, i interface{}, ResId int64) MSG {
-	var str string
-	if m, ok := i.(gjson.Result); ok {
-		if m.Type == gjson.JSON {
-			elem := bot.ConvertObjectMessage(m, true)
-			mid := bot.SendPrivateMessage(userId, &message.SendingMessage{Elements: elem})
-			if mid == -1 {
-				return Failed(100)
-			}
-			return OK(MSG{"message_id": mid})
-		}
-		str = func() string {
-			if m.Str != "" {
-				return m.Str
-			}
-			return m.Raw
-		}()
-	} else if s, ok := i.(string); ok {
-		str = s
-	}
-	if str == "" {
-		return Failed(100)
-	}
-	var elem []message.IMessageElement
-	println(str)
-	elem = append(elem, NewXmlMsg(str,ResId))
-
-	mid := bot.SendPrivateMessage(userId, &message.SendingMessage{Elements: elem})
-	if mid == -1 {
-		return Failed(100)
-	}
-	return OK(MSG{"message_id": mid})
-}
-
-func NewXmlMsg(template string,ResId int64) *message.ServiceElement{
-	var serviceid string
-	if ResId == 0{
-		serviceid ="2" //默认值2
-	}else{
-		serviceid = strconv.FormatInt(ResId,10)
-	}
-	//println(serviceid)
-	return &message.ServiceElement{
-		Id:      int32(ResId),
-		Content: template,
-		ResId:   serviceid,
-		SubType: "xml",
-	}
-}
-
-func (bot *CQBot) CQSendGroupMessage_XML(groupId int64, i interface{}, ResId int64) MSG {
-	var str string
-	fixAt := func(elem []message.IMessageElement) {
-		for _, e := range elem {
-			if at, ok := e.(*message.AtElement); ok && at.Target != 0 {
-				at.Display = "@" + func() string {
-					mem := bot.Client.FindGroup(groupId).FindMember(at.Target)
-					if mem != nil {
-						return mem.DisplayName()
-					}
-					return strconv.FormatInt(at.Target, 10)
-				}()
-			}
-		}
-	}
-	if m, ok := i.(gjson.Result); ok {
-		if m.Type == gjson.JSON {
-			elem := bot.ConvertObjectMessage(m, true)
-			fixAt(elem)
-			mid := bot.SendGroupMessage(groupId, &message.SendingMessage{Elements: elem})
-			if mid == -1 {
-				return Failed(100)
-			}
-			return OK(MSG{"message_id": mid})
-		}
-		str = func() string {
-			if m.Str != "" {
-				return m.Str
-			}
-			return m.Raw
-		}()
-	} else if s, ok := i.(string); ok {
-		str = s
-	}
-	if str == "" {
-		return Failed(100)
-	}
-	var elem []message.IMessageElement
-	println(str)
-	elem = append(elem, NewXmlMsg(str,ResId))
-	//fixAt(elem)
-	mid := bot.SendGroupMessage(groupId, &message.SendingMessage{Elements: elem})
-	if mid == -1 {
-		return Failed(100)
-	}
-	return OK(MSG{"message_id": mid})
-}
