+++ global/net.go
@@ -3,9 +3,11 @@ package global
 import (
 	"bytes"
 	"compress/gzip"
+	"github.com/Mrs4s/MiraiGo/message"
 	"github.com/tidwall/gjson"
 	"io/ioutil"
 	"net/http"
+	"strconv"
 	"strings"
 )
 
@@ -40,4 +42,20 @@ func QQMusicSongInfo(id string) (gjson.Result, error) {
 		return gjson.Result{}, err
 	}
 	return gjson.ParseBytes(d).Get("songinfo.data"), nil
+}
+
+func NewXmlMsg(template string,ResId int64) *message.ServiceElement{
+	var serviceid string
+	if ResId == 0{
+		serviceid ="2" //默认值2
+	}else{
+		serviceid = strconv.FormatInt(ResId,10)
+	}
+	//println(serviceid)
+	return &message.ServiceElement{
+		Id:      int32(ResId),
+		Content: template,
+		ResId:   serviceid,
+		SubType: "xml",
+	}
 }
